+++ app/src/main/java/ru/bank/app/command/GeneratorID.java
