+++ app/src/test/java/ru/bank/app/command/GeneratorIDTest.java
