+++ server/src/main/java/ru/bank/server/BankServer.java
@@ -3,13 +3,13 @@ package ru.bank.server;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
-import ru.bank.server.baseUsers.BaseUsers;
-import ru.bank.server.baseUsers.HisoryTransaction;
-import ru.bank.server.baseUsers.InfoTransaction;
 import ru.bank.server.command.PhonePayment;
 import ru.bank.server.exception.PaymentIndetifierException;
 import ru.bank.server.validation.ValidatorPaymentIndetifier;
 import ru.bank.users.User;
+import ru.bank.users.baseUsers.BaseUsers;
+import ru.bank.users.baseUsers.HisoryTransaction;
+import ru.bank.users.baseUsers.InfoTransaction;
 import ru.bank.users.paymentAttributes.Сurrency;
 
 import java.util.ArrayList;
