+++ server/src/main/java/ru/bank/server/command/Pay.java
