+++ server/src/main/java/ru/bank/server/command/PhonePayment.java
