+++ server/src/main/java/ru/bank/server/exception/PaymentIndetifierException.java
