+++ user/src/main/java/ru/bank/users/User.java
