+++ user/src/main/java/ru/bank/users/paymentAttributes/AccountIndividual.java
