+++ user/src/main/java/ru/bank/users/paymentAttributes/Сurrency.java
